+++ src/ptime.mli
@@ -46,7 +46,12 @@ module Span : sig
   (** {1:spans POSIX time spans} *)
 
   type t = span
-  (** The type for for signed, picosecond precision, POSIX time spans. *)
+  (** The type for signed, picosecond precision, POSIX time spans. *)
+
+  val v : int * int64 -> span
+  (** [v s] is like {!of_d_ps}[ s] but @raise Invalid_argument if
+      [s] is not in the right range. Use {!of_d_ps} to deal with
+      untrusted input. *)
 
   val zero : span
   (** [zero] is the neutral element of {!add}. *)
@@ -160,6 +165,11 @@ type t
 (** The type for picosecond precision POSIX timestamps in the range
     \[{!min};{!max}\]. Note that POSIX timestamps, and hence values of
     this type, are by definition always on the UTC timeline. *)
+
+val v : int * int64 -> t
+(** [v s] is [of_span (Span.v s)] but @raise Invalid_argument if [s]
+    is not in the right range. Use {!Span.of_d_ps} and {!of_span}
+    to deal with untrusted input. *)
 
 val epoch : t
 (** [epoch] is 1970-01-01 00:00:00 UTC. *)