+++ CHANGES.md
@@ -1,3 +1,6 @@
+
+* Relicense from BSD3 to ISC.
+
 v0.8.0 2015-12-24 Cambridge (UK)
 --------------------------------